+++ src/lib.rs
@@ -2,14 +2,10 @@
 #![warn(clippy::pedantic)]
 #![allow(clippy::missing_errors_doc, clippy::must_use_candidate, clippy::uninlined_format_args, clippy::doc_markdown)]
 
-mod comment_header;
-mod comment_list;
+mod codec;
 mod constants;
 mod decibels;
-mod discrete_comment_list;
 mod error;
-mod fixed_point_gain;
-mod opus_header;
 
 /// Functionality for escaping and unescaping values for command-line tools
 pub mod escaping;
@@ -27,19 +23,16 @@ pub mod interrupt;
 /// volume tags
 pub mod volume_rewrite;
 
-/// Functionality for determining BS.1770 loudness of Ogg Opus streams
-pub mod volume_analyzer;
+/// Functionality for manipulating headers
+pub mod header;
 
+/// Types for manipulating headers of Ogg Opus streams
+pub mod opus;
+
+/// Types for manipulating headers of Ogg Vorbis streams
+pub mod vorbis;
+
+pub use codec::*;
 pub use constants::global::*;
 pub use decibels::*;
 pub use error::*;
-
-/// Types for manipulating headers of Ogg Opus streams
-pub mod opus {
-    pub use crate::comment_header::*;
-    pub use crate::comment_list::*;
-    pub use crate::constants::opus::*;
-    pub use crate::discrete_comment_list::*;
-    pub use crate::fixed_point_gain::*;
-    pub use crate::opus_header::*;
-}