+++ src/volume_rewrite.rs
@@ -1,7 +1,8 @@
 use std::convert::{Into, TryFrom};
 
-use crate::header_rewriter::HeaderRewrite;
-use crate::opus::{CommentHeader, CommentList, FixedPointGain, OpusHeader, TAG_ALBUM_GAIN, TAG_TRACK_GAIN};
+use crate::header::{CommentList, FixedPointGain};
+use crate::header_rewriter::{CodecHeaders, HeaderRewrite, HeaderSummarize};
+use crate::opus::{TAG_ALBUM_GAIN, TAG_TRACK_GAIN};
 use crate::{Decibels, Error, R128_LUFS};
 
 /// Represents a target gain for an audio stream
@@ -76,6 +77,29 @@ pub struct OpusGains {
     pub album_r128: Option<Decibels>,
 }
 
+/// Returns the gains from the codec headers
+#[derive(Debug, Default)]
+pub struct GainsSummary {}
+
+impl HeaderSummarize for GainsSummary {
+    type Error = Error;
+    type Summary = OpusGains;
+
+    fn summarize(&self, headers: &CodecHeaders) -> Result<OpusGains, Error> {
+        match headers {
+            CodecHeaders::Opus(opus_header, comment_header) => {
+                let gains = OpusGains {
+                    output: opus_header.get_output_gain().into(),
+                    track_r128: comment_header.get_gain_from_tag(TAG_TRACK_GAIN).unwrap_or(None).map(Into::into),
+                    album_r128: comment_header.get_gain_from_tag(TAG_ALBUM_GAIN).unwrap_or(None).map(Into::into),
+                };
+                Ok(gains)
+            }
+            CodecHeaders::Vorbis(_, _) => Err(Error::UnsupportedCodec(headers.codec())),
+        }
+    }
+}
+
 /// Parameterization struct for `HeaderRewriter` to rewrite ouput gain and R128
 /// tags.
 #[derive(Debug)]
@@ -89,44 +113,41 @@ impl VolumeHeaderRewrite {
 
 impl HeaderRewrite for VolumeHeaderRewrite {
     type Error = Error;
-    type Summary = OpusGains;
 
-    fn summarize(&self, opus_header: &OpusHeader, comment_header: &CommentHeader) -> Result<OpusGains, Error> {
-        let gains = OpusGains {
-            output: opus_header.get_output_gain().into(),
-            track_r128: comment_header.get_gain_from_tag(TAG_TRACK_GAIN).unwrap_or(None).map(Into::into),
-            album_r128: comment_header.get_gain_from_tag(TAG_ALBUM_GAIN).unwrap_or(None).map(Into::into),
-        };
-        Ok(gains)
-    }
-
-    fn rewrite(&self, opus_header: &mut OpusHeader, comment_header: &mut CommentHeader) -> Result<(), Error> {
-        let new_header_gain = match self.config.output_gain {
-            VolumeTarget::ZeroGain => FixedPointGain::default(),
-            VolumeTarget::LUFS(target_lufs) => {
-                let volume_for_output_gain =
-                    self.config.volume_for_output_gain_calculation().expect("Precomputed volume unexpectedly missing");
-                FixedPointGain::try_from(target_lufs - volume_for_output_gain)?
+    fn rewrite(&self, headers: &mut CodecHeaders) -> Result<(), Error> {
+        match headers {
+            CodecHeaders::Opus(opus_header, comment_header) => {
+                let new_header_gain = match self.config.output_gain {
+                    VolumeTarget::ZeroGain => FixedPointGain::default(),
+                    VolumeTarget::LUFS(target_lufs) => {
+                        let volume_for_output_gain = self
+                            .config
+                            .volume_for_output_gain_calculation()
+                            .expect("Precomputed volume unexpectedly missing");
+                        FixedPointGain::try_from(target_lufs - volume_for_output_gain)?
+                    }
+                    VolumeTarget::NoChange => opus_header.get_output_gain(),
+                };
+                opus_header.set_output_gain(new_header_gain);
+                let compute_gain = |volume| -> Result<Option<FixedPointGain>, Error> {
+                    if let Some(volume) = volume {
+                        FixedPointGain::try_from(R128_LUFS - volume - new_header_gain.into()).map(Some)
+                    } else {
+                        Ok(None)
+                    }
+                };
+                let track_gain_r128 = compute_gain(self.config.track_volume)?;
+                let album_gain_r128 = compute_gain(self.config.album_volume)?;
+                for (tag, gain) in [(TAG_TRACK_GAIN, track_gain_r128), (TAG_ALBUM_GAIN, album_gain_r128)] {
+                    if let Some(gain) = gain {
+                        comment_header.set_tag_to_gain(tag, gain)?;
+                    } else {
+                        comment_header.remove_all(tag);
+                    }
+                }
+                Ok(())
             }
-            VolumeTarget::NoChange => opus_header.get_output_gain(),
-        };
-        opus_header.set_output_gain(new_header_gain);
-        let compute_gain = |volume| -> Result<Option<FixedPointGain>, Error> {
-            if let Some(volume) = volume {
-                FixedPointGain::try_from(R128_LUFS - volume - new_header_gain.into()).map(Some)
-            } else {
-                Ok(None)
-            }
-        };
-        let track_gain_r128 = compute_gain(self.config.track_volume)?;
-        let album_gain_r128 = compute_gain(self.config.album_volume)?;
-        for (tag, gain) in [(TAG_TRACK_GAIN, track_gain_r128), (TAG_ALBUM_GAIN, album_gain_r128)] {
-            if let Some(gain) = gain {
-                comment_header.set_tag_to_gain(tag, gain)?;
-            } else {
-                comment_header.remove_all(tag);
-            }
+            CodecHeaders::Vorbis(_, _) => Err(Error::UnsupportedCodec(headers.codec())),
         }
-        Ok(())
     }
 }