+++ src/vorbis/mod.rs
@@ -1,0 +1,5 @@
+mod comment_header;
+mod id_header;
+
+pub use comment_header::{CommentHeader, Specifics as CommentHeaderSpecifics};
+pub use id_header::*;